requirements.txt
authlib==1.5.2
Flask==3.1.0
Flask-Session==0.8.0
google_api_python_client==2.167.0
gspread==6.2.0
Jinja2==3.1.6
oauth2client==4.1.3
playwright==1.52.0
protobuf==6.30.2
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
gunicorn==23.0.0